Refactored Format class, now in formats.py
