+++ pywps/inout.py
@@ -1,6 +1,7 @@
 from abc import ABCMeta, abstractmethod, abstractproperty
 from pywps._compat import text_type, StringIO
 from pywps.formats import FORMATS
+from pywps.formats import Format
 import tempfile
 
 LITERAL_DATA_TYPES = ['string', 'float', 'integer', 'boolean']
@@ -10,95 +11,6 @@ class SOURCE_TYPE:
     FILE = 1
     STREAM = 2
     DATA = 3
-
-
-class FormatBase(object):
-    """Input/output format specification
-    """
-    def __init__(self, mime_type, schema=None, encoding=None, validator=None):
-
-        self._mime_type = None
-        self._encoding = None
-        self._schema = None
-        self._validator = None
-
-        self.mime_type = mime_type
-        self.encoding = encoding
-        self.schema = schema
-        self.validator = validator
-
-
-    @property
-    def mime_type(self):
-        """Get format mime type
-        :rtype: String
-        """
-
-        if self._mime_type in FORMATS:
-            return FORMATS[self._mime_type][0]
-        else:
-            return self._mime_type
-
-    @mime_type.setter
-    def mime_type(self, mime_type):
-        """Set format mime type
-        """
-
-        self._mime_type = mime_type
-
-    def get_extension(self):
-        return self._mime_type[1]
-
-    @property
-    def encoding(self):
-        """Get format encoding
-        :rtype: String
-        """
-
-        if self._encoding:
-            return self._encoding
-        else:
-            return ''
-
-    @encoding.setter
-    def encoding(self, encoding):
-        """Set format encoding
-        """
-
-        self._encoding = encoding
-
-    @property
-    def schema(self):
-        """Get format schema
-        :rtype: String
-        """
-        if self._schema:
-            return self._schema
-        else:
-            return ''
-
-    @schema.setter
-    def schema(self, schema):
-        """Set format schema
-        """
-        self._schema = schema
-
-    @property
-    def validator(self):
-        """Get format validator
-        :rtype: String
-        """
-        if self._validator:
-            return self._validator
-        else:
-            return ''
-
-    @validator.setter
-    def validator(self, validator):
-        """Set format validator
-        """
-        self._validator = validator
-
 
 class DataTypeAbstract(object):
     """LiteralObject data_type abstract class
